Move a couple more things out of commands
